feat: change logo, move css to mui
