+++ ui/hockey-teeth/src/layouts/HomeLayout.tsx
@@ -1,16 +1,41 @@
 import React from 'react';
 import { Outlet } from 'react-router-dom';
-import { Container } from '@mui/material';
-import logo from '../assets/logo.png';
+import {
+  Container, Box, AppBar, Toolbar,
+} from '@mui/material';
+import logo from '../assets/logo.svg';
 import Footer from '../components/Footer';
 import MenuBar from '../components/MenuBar';
 
 function HomeLayout() {
   return (
     <div className="App">
-      <header className="App-header">
-        <img src={logo} className="App-logo" alt="logo" />
-      </header>
+      <AppBar
+        position="static"
+        color="transparent"
+        elevation={0}
+        sx={{ bgcolor: 'transparent' }}
+      >
+        <Toolbar
+          sx={{
+            display: 'flex',
+            flexDirection: 'column',
+            alignItems: 'center',
+            justifyContent: 'center',
+          }}
+        >
+          <Box
+            component="img"
+            src={logo}
+            alt="logo"
+            sx={{
+              height: '150px',
+              width: '375px',
+              pointerEvents: 'none',
+            }}
+          />
+        </Toolbar>
+      </AppBar>
       <Container maxWidth="md">
         <MenuBar />
         <Outlet />
